pyproject.toml
[project]
name = "document-scanner-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.52.0",
    "lxml>=6.0.2",
    "mistralai>=1.2.0",
    "openai>=2.8.1",
    "opencv-python-headless>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "protobuf>=6.33.1",
    "pypdf>=6.4.0",
    "pytesseract>=0.3.13",
    "pytest>=9.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
